Modify link text
As per a suggestion from Stirling I changed the link text for the link to
the sender of the page to make it a little more clear what is going on.

A better solution would be to look up the sender in the user's contacts
list and use the label from that.