Allow tag register from anon bump
I put an AJAX tag register form in the Explain dialog and gave it a call
back to seet your "own_tag" cookie and then jump to the no longer
anonymous bump URL.

This would seem to allow for a much smoother interaction: Someone sends
you an "anon bump" URL, you follow it, click on explain, enter your own
"avatar" URL and tag it, and you're automatically presented with a
personalized version of the previously anonymous bump URL and you're
participating in the network.

The user need never go the the '/register' page.